requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
crispy-bootstrap4==2023.1
cryptography==41.0.4
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.6
django-allauth==0.57.0
django-crispy-forms==2.1
django-hidefield==0.1.2
django-jazzmin==2.6.0
django-tinymce==3.6.1
environs==9.5.0
fontawesomefree==6.4.2
idna==3.4
marshmallow==3.20.1
oauthlib==3.2.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
